refactor(services): R-02..R-05 — extract AutoStartToggle, ApiKeyField, CliproxyConnectionPanel, NinerouterEmbedFrame
Extract 4 shared UI components from inline definitions in CliproxyServiceTab and
NinerouterServiceTab, eliminating duplication. Pure refactor — no behavior changes.
CliproxyServiceTab: 199 → 26 lines; NinerouterServiceTab: 193 → 43 lines.
